@@ -1,10 +1,10 @@
 typedef ValueBuilder = String Function(dynamic value);
 
 class JsonTableColumn {
-  String label;
-  String field;
-  String defaultValue;
-  ValueBuilder valueBuilder;
+  String? label;
+  String? field;
+  String? defaultValue;
+  ValueBuilder? valueBuilder;
 
   JsonTableColumn(this.field,
       {this.label, this.defaultValue = '', this.valueBuilder}) {
